// src/api/back/models/TournamentBracketParticipantDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Public REST api for dota2classic
 * All stuff
 *
 * The version of the OpenAPI document: 1.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from "../runtime";
import {
  PlayerPreviewDto,
  PlayerPreviewDtoFromJSON,
  PlayerPreviewDtoFromJSONTyped,
  PlayerPreviewDtoToJSON,
  TeamDto,
  TeamDtoFromJSON,
  TeamDtoFromJSONTyped,
  TeamDtoToJSON
} from "./";

/**
 *
 * @export
 * @interface TournamentBracketParticipantDto
 */
export interface TournamentBracketParticipantDto {
  /**
   *
   * @type {number}
   * @memberof TournamentBracketParticipantDto
   */
  id: number;
  /**
   *
   * @type {number}
   * @memberof TournamentBracketParticipantDto
   */
  tournamentId: number;
  /**
   *
   * @type {PlayerPreviewDto}
   * @memberof TournamentBracketParticipantDto
   */
  profile?: PlayerPreviewDto;
  /**
   *
   * @type {TeamDto}
   * @memberof TournamentBracketParticipantDto
   */
  team?: TeamDto;
}

export function TournamentBracketParticipantDtoFromJSON(json: any): TournamentBracketParticipantDto {
  return TournamentBracketParticipantDtoFromJSONTyped(json, false);
}

export function TournamentBracketParticipantDtoFromJSONTyped(
  json: any,
  ignoreDiscriminator: boolean
): TournamentBracketParticipantDto {
  if (json === undefined || json === null) {
    return json;
  }
  return {
    id: json["id"],
    tournamentId: json["tournament_id"],
    profile: !exists(json, "profile") ? undefined : PlayerPreviewDtoFromJSON(json["profile"]),
    team: !exists(json, "team") ? undefined : TeamDtoFromJSON(json["team"])
  };
}

export function TournamentBracketParticipantDtoToJSON(value?: TournamentBracketParticipantDto | null): any {
  if (value === undefined) {
    return undefined;
  }
  if (value === null) {
    return null;
  }
  return {
    id: value.id,
    tournament_id: value.tournamentId,
    profile: PlayerPreviewDtoToJSON(value.profile),
    team: TeamDtoToJSON(value.team)
  };
}